Added Jest for server
