+++ server/package.json
@@ -4,11 +4,15 @@
   "main": "src/index.js",
   "license": "MIT",
   "scripts": {
-    "start": "nodemon --legacy-watch src/index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node src/index.js",
+    "dev": "nodemon --legacy-watch src/index.js",
+    "test": "jest",
+    "coverage": "jest --collectCoverageFrom=src/**.js --coverage ."
   },
   "devDependencies": {
-    "nodemon": "^1.18.4"
+    "jest": "^23.5.0",
+    "nodemon": "^1.18.4",
+    "supertest": "^3.1.0"
   },
   "dependencies": {
     "body-parser": "^1.18.3",
